checkpoint commit so I can circle back to other issues.
