+++ pkg/registrars/registration/decoy-registrar.go
@@ -8,10 +8,3 @@ import (
 func NewDecoyRegistrar() *dr.DecoyRegistrar {
 	return dr.NewDecoyRegistrar()
 }
-
-// NewDecoyRegistrarWithDialer returns a decoy registrar with custom dialer.
-//
-// Deprecated: Set dialer in tapdace.Dialer.DialerWithLaddr instead.
-func NewDecoyRegistrarWithDialer(dialer dr.DialFunc) *dr.DecoyRegistrar {
-	return dr.NewDecoyRegistrarWithDialer(dialer)
-}
