+++ pkg/station/config_test.go
@@ -1,4 +1,4 @@
-package lib
+package station
 
 import (
 	"os"
