+++ pkg/transports/wrapping/prefix/prefix_test.go
@@ -8,6 +8,7 @@ import (
 	"io"
 	"os"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/require"
 	"golang.org/x/crypto/curve25519"
@@ -64,11 +65,10 @@ func TestSuccessfulWrap(t *testing.T) {
 				require.Nil(t, err)
 
 				var buf [4096]byte
-				var buffer bytes.Buffer
 				n, _ := sfp.Read(buf[:])
-				buffer.Write(buf[:n])
+				buffer := bytes.NewBuffer(buf[:n])
 
-				_, wrapped, err := transport.WrapConnection(&buffer, sfp, reg.PhantomIp, manager)
+				_, wrapped, err := transport.WrapConnection(buffer, sfp, reg.PhantomIp, manager)
 				if id != idx {
 					require.ErrorIs(t, err, ErrIncorrectPrefix)
 					continue
@@ -399,3 +399,65 @@ func TestSuccessfulWrapBase64(t *testing.T) {
 
 }
 */
+
+// Test End to End client WrapConn to Server WrapConnection
+func TestPrefixEndToEnd(t *testing.T) {
+	testSubnetPath := conjurepath.Root + "/pkg/station/lib/test/phantom_subnets.toml"
+	os.Setenv("PHANTOM_SUBNET_LOCATION", testSubnetPath)
+
+	_, private, _ := ed25519.GenerateKey(rand.Reader)
+
+	var curve25519Public, curve25519Private [32]byte
+	extra25519.PrivateKeyToCurve25519(&curve25519Private, private)
+	curve25519.ScalarBaseMult(&curve25519Public, &curve25519Private)
+
+	var transport = Transport{
+		TagObfuscator:     transports.CTRObfuscator{},
+		Privkey:           curve25519Private,
+		SupportedPrefixes: defaultPrefixes,
+	}
+	message := []byte(`test message!`)
+
+	for idx := range defaultPrefixes {
+
+		var p int32 = int32(idx)
+		params := &pb.PrefixTransportParams{PrefixId: &p}
+		manager := tests.SetupRegistrationManager(tests.Transport{Index: pb.TransportType_Prefix, Transport: transport})
+		c2p, sfp, reg := tests.SetupPhantomConnections(manager, pb.TransportType_Prefix, params, randomizeDstPortMinVersion)
+		defer c2p.Close()
+		defer sfp.Close()
+		require.NotNil(t, reg)
+
+		go func() {
+			var buf [4096]byte
+			n, _ := sfp.Read(buf[:])
+			buffer := bytes.NewBuffer(buf[:n])
+
+			_, c, err := transport.WrapConnection(buffer, sfp, reg.PhantomIp, manager)
+			require.Nil(t, err, "error getting wrapped connection")
+
+			received := make([]byte, len(message))
+			_, err = io.ReadFull(c, received)
+			require.Nil(t, err, "failed reading from connection")
+			_, err = c.Write(received)
+			require.Nil(t, err, "failed writing to connection")
+		}()
+
+		clientPrefix, err := TryFromID(PrefixID(p))
+		require.Nil(t, err)
+		ClientTransport := &ClientTransport{Prefix: clientPrefix, parameters: params}
+		err = ClientTransport.PrepareKeys(curve25519Public, reg.Keys.SharedSecret, reg.Keys.TransportReader)
+		clientConn, err := ClientTransport.WrapConn(c2p)
+		require.Nil(t, err, "error getting wrapped connection")
+
+		err = clientConn.SetDeadline(time.Now().Add(3 * time.Second))
+		require.Nil(t, err, "error setting deadline")
+
+		_, err = clientConn.Write(message)
+		require.Nil(t, err, "failed writing to connection")
+		cbuf := make([]byte, len(message))
+		_, err = io.ReadFull(clientConn, cbuf)
+		require.Nil(t, err, "failed reading from connection")
+		require.True(t, bytes.Equal(message, cbuf), "%s\n%s", string(message), string(cbuf))
+	}
+}
